commented out some listings fn and order history is now displaying data but not all data
